refactor(providers): move reusable OpenAI request schemas to decorator-free modules
OpenAI-compatible providers (google, ollama, deepseek, openrouter,
nvidia_nim, groq) reuse OpenAI request schemas by importing them from
`openai.chat` / `openai.audio` -- the same files that carry the
`@register` endpoint decorators. The dir->file collapse merged the old
decorator-free `models.py` into those endpoint files, so touching a
compatible provider's lazy export (or resolving its config LazyType)
imported and registered the OpenAI chat/audio endpoints as a side
effect, breaking the documented import-laziness contract.

Extract the reusable request schemas (chat: OpenAIChatCompletionsRequest
and its ~20 sub-models; audio: AudioSpeechRequest, AudioTranscriptionRequest)
into decorator-free `openai/_chat_schemas.py` and `openai/_audio_schemas.py`.
`openai/chat.py` and `openai/audio.py` re-import them for the layer-2 path;
the compatible providers and every request-schema config LazyType now point
at the schema modules instead of the endpoint modules.

A comprehensive per-provider probe confirms the full contract: touching
any of the 14 providers' lazy exports loads none of the 31 endpoint
modules belonging to another provider. Importer test suites (providers,
connections, imodel, operations, session, docs) pass: 1437.
