+++ lionagi/providers/google/_config.py
@@ -12,7 +12,7 @@ class GeminiChatConfigs(ProviderConfig, Enum):
         "chat/completions",
         ["chat"],
         EndpointType.API,
-        LazyType("lionagi.providers.openai.chat:OpenAIChatCompletionsRequest"),
+        LazyType("lionagi.providers.openai._chat_schemas:OpenAIChatCompletionsRequest"),
         "https://generativelanguage.googleapis.com/v1beta/openai",
         "bearer",
     )
