+++ lionagi/providers/openai/_config.py
@@ -12,7 +12,7 @@ class OpenAIConfigs(ProviderConfig, Enum):
         "chat/completions",
         ["chat"],
         EndpointType.API,
-        LazyType("lionagi.providers.openai.chat:OpenAIChatCompletionsRequest"),
+        LazyType("lionagi.providers.openai._chat_schemas:OpenAIChatCompletionsRequest"),
         "https://api.openai.com/v1",
         "bearer",
     )
@@ -36,7 +36,7 @@ class OpenAIConfigs(ProviderConfig, Enum):
         "audio/speech",
         ["tts"],
         EndpointType.API,
-        LazyType("lionagi.providers.openai.audio:AudioSpeechRequest"),
+        LazyType("lionagi.providers.openai._audio_schemas:AudioSpeechRequest"),
         "https://api.openai.com/v1",
         "bearer",
     )
@@ -44,7 +44,7 @@ class OpenAIConfigs(ProviderConfig, Enum):
         "audio/transcriptions",
         ["stt", "whisper"],
         EndpointType.API,
-        LazyType("lionagi.providers.openai.audio:AudioTranscriptionRequest"),
+        LazyType("lionagi.providers.openai._audio_schemas:AudioTranscriptionRequest"),
         "https://api.openai.com/v1",
         "bearer",
     )
